+++ Backend/calculation_engine/src/main/java/com/wisetech/rras/calculationengine/domain/LoanExposureSnapshot.java
@@ -22,7 +22,7 @@ public class LoanExposureSnapshot {
 
     @Id
     @Column(name = "snapshot_id")
-    private Long snapshotId;
+    private int snapshotId;
 
     @Id
     @Column(name = "loan_id")
