+++ Backend/calculation_engine/src/main/java/com/wisetech/rras/calculationengine/domain/LoanExposureSnapshotId.java
@@ -10,6 +10,6 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 public class LoanExposureSnapshotId implements Serializable {
-    private Long snapshotId;
+    private int snapshotId;
     private Long loanId;
 }
