+++ Backend/calculation_engine/src/main/java/com/wisetech/rras/calculationengine/domain/MetricComponent.java
@@ -19,7 +19,7 @@ public class MetricComponent {
 
     @Id
     @Column(name = "snapshot_id")
-    private Long snapshotId;
+    private int snapshotId;
 
     @Id
     @Column(name = "loan_id")
